Generate Rust type nodes
This PR generates 21 of the 28 `type`-category nodes from `@codama/spec`, completing first-class generation support for all 8 spec categories. The 7 nestable wrappers (`fixedSize`, `sizePrefix`, `pre/postOffset`, `sentinel`, `hiddenPrefix/Suffix`) and the four type-category unions stay hand-written, since they carry generic variant types, a cross-category Link member, and bespoke `TryFrom` bridges that don't fit the standard renderers — and they're the natural deletion boundary if a future spec drops nested unions.

The generator gains a `NodeMacroFlavor` classifier (`nodeFlavor.ts`) that picks the emission macro per node from the spec: `nestedTypeNode.wrappers` are skipped, `enumVariantTypeNode` members and `structFieldTypeNode` get `#[node]`, and other type nodes get `#[type_node]`. `CategoryRouting` gains an `emitUnions` flag so `type` skips its bespoke unions, and `struct` joins the Rust keyword set so `enumStructVariantTypeNode.struct` renders as `r#struct`.

Reconciliation is mostly contained to `codama-nodes/src/type_nodes/`. Spec-driven type changes: `amountTypeNode.decimals → u32`, the enum-variant `discriminator → Option<u32>`, and `optionTypeNode.fixed → Option<bool>`. The box-all-union rule boxes `array/map/set.count → Box<CountNode>` and `structFieldTypeNode.{type, default_value}`, which cascades into a few `codama-attributes`/`codama-koroks`/`codama-korok-visitors` files and their tests. The hand-written nestable nodes are left untouched.

1008 cargo tests pass unchanged; `pnpm test` grew by 8 to 148; fmt, clippy, and the `pnpm generate` round-trip stay clean. The override surface is unchanged.

+++ codama-korok-visitors/src/combine_types_visitor.rs
@@ -240,7 +240,7 @@ impl KorokVisitor for CombineTypesVisitor {
             .ast
             .discriminant
             .as_ref()
-            .and_then(|(_, x)| x.as_unsigned_integer::<usize>().ok());
+            .and_then(|(_, x)| x.as_unsigned_integer::<u32>().ok());
 
         korok.node = match korok.ast.fields {
             syn::Fields::Named(_) => {
